requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-countries==7.5.1
django-filter==23.3
django-phone-auth==0.3.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Markdown==3.5
oauthlib==3.2.2
phonenumbers==8.13.22
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
PythonTurtle==0.3.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tk==0.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.6.0
